Move relationships() to DataSources
